Oplossingen/labo7-modeloplossing-preview/requirements.txt
beautifulsoup4==4.10.0
Jinja2==3.1.1
MarkupSafe==2.1.1
mistune==2.0.2
PyYAML==6.0
soupsieve==2.3.1
